chore: address comments
