+++ packages/server/src/types/validator/methods.ts
@@ -18,22 +18,11 @@
  *
  */
 
-import { validateParam } from './utils';
+export type IMethodValidation = {
+  [index: number]: IMethodParamSchema;
+};
 
-import { IMethodValidation } from '../types/validator/methods';
-
-export function validateParams(params: any[], indexes: IMethodValidation) {
-  for (const index of Object.keys(indexes)) {
-    const validation = indexes[Number(index)];
-    const param = params[Number(index)];
-
-    validateParam(index, param, validation);
-  }
-}
-
-export * from './constants';
-export * from './types';
-export * from './objectTypes';
-export * from './utils';
-export * from './methods';
-export * as Validator from '.';
+export type IMethodParamSchema = {
+  type: string;
+  required?: boolean;
+};
